# src/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "2.0.0",
    "title": "WPPConnect API Rest",
    "description": "Welcome to the wppconnect-server API documentation. This API provides a set of endpoints to interact with the wppconnect-server application, allowing you to build integrations and automate interactions with WhatsApp."
  },
...
  "paths": {
...
    "/api/{secretkey}/show-all-sessions": {
      "get": {
        "tags": ["Auth"],
        "description": "",
        "operationId": "showAllSessions",
        "parameters": [
          {
            "name": "secretkey",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "example": "THISISMYSECURETOKEN"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/{secretkey}/start-all": {
      "post": {
        "tags": ["Auth"],
        "description": "",
        "operationId": "startAllSessions",
        "parameters": [
          {
            "name": "secretkey",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "example": "THISISMYSECURECODE"
            }
          },
          {
            "name": "session",
            "schema": {
              "type": "string",
              "example": "NERDWHATS_AMERICA"
            },
            "in": "query"
          },
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/{session}/check-connection-session": {
      "get": {
        "tags": ["Auth"],
        "description": "",
        "operationId": "CheckConnectionState",
        "parameters": [
          {
            "name": "session",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "example": "NERDWHATS_AMERICA"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/{session}/get-media-by-message/{messageId}": {
      "get": {
        "tags": ["Messages"],
        "description": "",
        "operationId": "getMediaByMessage",
        "parameters": [
          {
            "name": "session",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "example": "messageId"
            }
          },
          {
            "name": "messageId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/api/{session}/qrcode-session": {
      "get": {
        "tags": ["Auth"],
        "description": "",
        "operationId": "getQrCode",
        "parameters": [
          {
            "name": "session",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "example": "NERDWHATS_AMERICA"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/{session}/start-session": {
      "post": {
        "tags": ["Auth"],
        "description": "",
        "operationId": "startSession",
        "parameters": [
          {
            "name": "session",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "example": "NERDWHATS_AMERICA"
            }
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "webhook": {
                    "type": "string"
                  },
                  "waitQrCode": {
                    "type": "boolean"
                  },
                  "proxyUrl": {
                    "type": "string"
                  }
                }
              },
              "example": {
                "webhook": "",
                "waitQrCode": false,
                "proxyUrl": ""
              }
            }
          }
        }
      }
    },
    "/api/{session}/logout-session": {
      "post": {
        "tags": ["Auth"],
        "description": "This route logout and delete session data",
        "operationId": "logoutSession",
        "parameters": [
          {
            "name": "session",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "example": "NERDWHATS_AMERICA"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/api/{session}/close-session": {
      "post": {
        "tags": ["Auth"],
        "description": "",
        "operationId": "closeSession",
        "parameters": [
          {
            "name": "session",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "example": "NERDWHATS_AMERICA"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/{session}/subscribe-presence": {
      "post": {
        "tags": ["Misc"],
        "description": "",
        "operationId": "subscribePresence",
        "parameters": [
          {
            "name": "session",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "example": "NERDWHATS_AMERICA"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "phone": {
                    "type": "string"
                  },
                  "isGroup": {
                    "type": "boolean"
                  },
                  "all": {
                    "type": "boolean"
                  }
                }
              },
              "example": {
                "phone": "5521999999999",
                "isGroup": false,
                "all": false
              }
            }
          }
        }
      }
    },
    "/api/{session}/download-media": {
      "post": {
        "tags": ["Messages"],
        "description": "",
        "operationId": "downloadMediabyMessage",
        "parameters": [
          {
            "name": "session",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "example": "NERDWHATS_AMERICA"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "messageId": {
                    "type": "string"
                  }
                }
              },
              "example": {
                "messageId": "<messageId>"
              }
            }
          }
        }
      }
    },
...
    "/api/{session}/status-session": {
      "get": {
        "tags": ["Auth"],
        "description": "",
        "operationId": "closeSession",
        "parameters": [
          {
            "name": "session",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "example": "NERDWHATS_AMERICA"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/api/{session}/edit-business-profile": {
      "post": {
        "tags": ["Profile"],
        "description": "Edit your bussiness profile",
        "operationId": "editBusinessProfile",
        "parameters": [
          {
            "name": "session",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "example": "NERDWHATS_AMERICA"
            }
          },
          {
            "name": "obj",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "adress": {
                  "type": "string",
                  "example": "Av. Nossa Senhora de Copacabana, 315"
                },
                "email": {
                  "type": "string",
                  "example": "test@test.com.br"
                },
                "categories": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "example": "133436743388217"
                    },
                    "localized_display_name": {
                      "type": "string",
                      "example": "Artes e entretenimento"
                    },
                    "not_a_biz": {
                      "type": "boolean",
                      "example": false
                    }
                  },
                  "required": ["id", "localized_display_name", "not_a_biz"]
                },
                "website": {
                  "type": "array",
                  "example": [
                    "https://www.wppconnect.io",
                    "https://www.teste2.com.br"
                  ],
                  "items": {
                    "type": "string"
                  }
                }
              },
              "required": ["adress", "email", "categories", "website"]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "adress": {
                    "type": "string"
                  },
                  "email": {
                    "type": "string"
                  },
                  "categories": {
                    "type": "object"
                  },
                  "websites": {
                    "type": "array"
                  }
                }
              },
              "example": {
                "adress": "Av. Nossa Senhora de Copacabana, 315",
                "email": "test@test.com.br",
                "categories": {
                  "$id": "133436743388217",
                  "$localized_display_name": "Artes e entretenimento",
                  "$not_a_biz": false
                },
                "website": [
                  "https://www.wppconnect.io",
                  "https://www.teste2.com.br"
                ]
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
...
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}